docker/notifier/requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
jinja2==3.1.4
